backend/requirements.txt
fastapi 
uvicorn 
pandas 
requests
sqlalchemy
selenium
webdriver-manager
python-dotenv
psycopg2
alembic